Add CellStateDropState and fancy drag and drop support
