+++ ui-cells/src/jbMain/kotlin/com/alexvanyo/composelife/ui/cells/SelectionHandle.kt
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alexvanyo.composelife.ui.cells
+
+import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+/**
+ * A selection handle.
+ */
+@Composable
+fun SelectionHandle(
+    isActive: Boolean,
+    modifier: Modifier = Modifier,
+    selectionColor: Color = MaterialTheme.colorScheme.secondary,
+) {
+    Box(
+        modifier = modifier.size(48.dp),
+        contentAlignment = Alignment.Center,
+    ) {
+        val size by animateDpAsState(if (isActive) 24.dp else 16.dp)
+        val elevation by animateDpAsState(if (isActive) 4.dp else 0.dp)
+        Surface(
+            modifier = Modifier.size(size),
+            shape = CircleShape,
+            shadowElevation = elevation,
+            color = selectionColor,
+        ) {}
+    }
+}
