+++ ui-cells/src/jbMain/kotlin/com/alexvanyo/composelife/ui/cells/SelectionOverlay.kt
@@ -16,95 +16,50 @@
 
 package com.alexvanyo.composelife.ui.cells
 
-import androidx.annotation.FloatRange
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.AnimationVector2D
-import androidx.compose.animation.core.VectorConverter
-import androidx.compose.animation.core.VisibilityThreshold
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.core.spring
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.interaction.collectIsDraggedAsState
-import androidx.compose.foundation.interaction.collectIsHoveredAsState
-import androidx.compose.foundation.interaction.collectIsPressedAsState
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.requiredSize
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.Stable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.saveable.Saver
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Rect
 import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.geometry.toRect
-import androidx.compose.ui.graphics.BlendMode
+import androidx.compose.ui.geometry.center
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.graphics.PathEffect
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.drawscope.DrawScope
-import androidx.compose.ui.graphics.drawscope.DrawScope.Companion.DefaultBlendMode
-import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.graphics.drawscope.ContentDrawScope
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.layout
 import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.Constraints
 import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.round
+import androidx.compose.ui.unit.toIntRect
 import androidx.compose.ui.unit.toOffset
-import androidx.compose.ui.util.packInts
-import androidx.compose.ui.util.unpackInt1
-import androidx.compose.ui.util.unpackInt2
+import androidx.compose.ui.unit.toRect
+import androidx.compose.ui.unit.toSize
+import com.alexvanyo.composelife.geometry.times
 import com.alexvanyo.composelife.model.CellState
 import com.alexvanyo.composelife.model.CellWindow
 import com.alexvanyo.composelife.model.di.CellStateParserProvider
-import com.alexvanyo.composelife.parameterizedstring.parameterizedStringResolver
 import com.alexvanyo.composelife.sessionvalue.SessionValue
 import com.alexvanyo.composelife.sessionvalue.preLocalSessionId
 import com.alexvanyo.composelife.sessionvalue.rememberSessionValueHolder
-import com.alexvanyo.composelife.ui.cells.resources.SelectingBoxHandle
-import com.alexvanyo.composelife.ui.cells.resources.Strings
 import com.alexvanyo.composelife.ui.util.AnchoredDraggable2DState
 import com.alexvanyo.composelife.ui.util.AnimatedContent
-import com.alexvanyo.composelife.ui.util.DraggableAnchors2D
 import com.alexvanyo.composelife.ui.util.TargetState
-import com.alexvanyo.composelife.ui.util.anchoredDraggable2D
-import com.alexvanyo.composelife.ui.util.snapTo
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.launch
 import kotlin.math.max
 import kotlin.math.min
 import kotlin.math.roundToInt
+import kotlin.uuid.Uuid
 
 /**
- * The overlay based on the [selectionState].
+ * The overlay based on the [selectionSessionState].
  */
 context(CellStateParserProvider)
-@Suppress("LongMethod", "LongParameterList")
+@Suppress("LongMethod", "LongParameterList", "CyclomaticComplexMethod")
 @Composable
 fun SelectionOverlay(
     selectionSessionState: SessionValue<SelectionState>,
@@ -112,6 +67,7 @@ fun SelectionOverlay(
     getSelectionCellState: (SelectionState) -> CellState,
     scaledCellDpSize: Dp,
     cellWindow: CellWindow,
+    pixelOffsetFromCenter: Offset,
     modifier: Modifier = Modifier,
 ) {
     val selectionSessionStateValueHolder = rememberSessionValueHolder(
@@ -122,6 +78,26 @@ fun SelectionOverlay(
         valueSaver = SelectionState.Saver,
     )
     val sessionValue = selectionSessionStateValueHolder.sessionValue
+
+    val scaledCellPixelSize = with(LocalDensity.current) { scaledCellDpSize.toPx() }
+    val cellStateDropStateHolder = rememberMutableCellStateDropStateHolder { dropOffset, cellState ->
+        setSelectionSessionState(
+            SessionValue(
+                sessionId = Uuid.random(),
+                valueId = Uuid.random(),
+                value = SelectionState.Selection(
+                    cellState = cellState,
+                    offset = (
+                        cellWindow.topLeft.toOffset() + (dropOffset / scaledCellPixelSize) -
+                            cellState.boundingBox.size.toSize().center
+                        ).round(),
+                ),
+            ),
+        )
+    }
+
+    val dropAvailableBorderColor = MaterialTheme.colorScheme.tertiary
+    val dropPreviewCellStateBorderColor = MaterialTheme.colorScheme.secondary
 
     AnimatedContent(
         targetState = TargetState.Single(
@@ -136,27 +112,59 @@ fun SelectionOverlay(
             } to preLocalSessionId
         },
         modifier = modifier
-            .cellStateDragAndDropTarget { cellState ->
-                val boundingBoxSize = cellState.boundingBox.size
-                selectionSessionStateValueHolder.setValue(
-                    SelectionState.Selection(
-                        cellState = cellState,
-                        offset = (
-                            cellWindow.center.toOffset() -
-                                Offset(boundingBoxSize.width - 1f, boundingBoxSize.height - 1f) / 2f
-                            ).round(),
-                    ),
-                )
+            .drawWithContent {
+                drawContent()
+
+                when (cellStateDropStateHolder.cellStateDropState) {
+                    CellStateDropState.ApplicableDropAvailable,
+                    is CellStateDropState.DropPreview,
+                    -> {
+                        drawDashedRect(
+                            selectionColor = dropAvailableBorderColor,
+                            strokeWidth = 4.dp.toPx(),
+                            intervals = floatArrayOf(
+                                24.dp.toPx(),
+                                24.dp.toPx(),
+                            ),
+                            phase = 12.dp.toPx(),
+                        )
+                    }
+                    CellStateDropState.None -> Unit
+                }
+            }
+            .graphicsLayer {
+                this.translationX = -pixelOffsetFromCenter.x
+                this.translationY = -pixelOffsetFromCenter.y
             }
             .requiredSize(
                 scaledCellDpSize * cellWindow.width,
                 scaledCellDpSize * cellWindow.height,
-            ),
+            )
+            .cellStateDragAndDropTarget(
+                mutableCellStateDropStateHolder = cellStateDropStateHolder,
+            )
+            .drawWithContent {
+                drawContent()
+
+                when (val cellStateDropState = cellStateDropStateHolder.cellStateDropState) {
+                    CellStateDropState.ApplicableDropAvailable -> Unit
+                    is CellStateDropState.DropPreview -> {
+                        drawDropPreview(
+                            dropPreview = cellStateDropState,
+                            scaledCellDpSize = scaledCellDpSize,
+                            cellStateOutlineColor = dropPreviewCellStateBorderColor,
+                        )
+                    }
+
+                    CellStateDropState.None -> Unit
+                }
+            },
     ) { (targetSelectionSessionState, _) ->
         when (val targetSelectionState = targetSelectionSessionState.value) {
             SelectionState.NoSelection -> {
                 Spacer(Modifier.fillMaxSize())
             }
+
             is SelectionState.SelectingBox.FixedSelectingBox -> {
                 @Suppress("UNCHECKED_CAST")
                 (
@@ -173,6 +181,7 @@ fun SelectionOverlay(
                     )
                     )
             }
+
             is SelectionState.SelectingBox.TransientSelectingBox -> {
                 TransientSelectingBoxOverlay(
                     selectionState = targetSelectionState,
@@ -181,426 +190,44 @@ fun SelectionOverlay(
                     modifier = Modifier.fillMaxSize(),
                 )
             }
+
             is SelectionState.Selection -> {
                 @Suppress("UNCHECKED_CAST")
-                (
-                    SelectionBoxOverlay(
-                        selectionSessionState = targetSelectionSessionState as SessionValue<SelectionState.Selection>,
-                        setSelectionState = selectionSessionStateValueHolder::setValue,
-                        getSelectionCellState = {
-                            getSelectionCellState(targetSelectionState)
-                        },
-                        scaledCellPixelSize = with(LocalDensity.current) { scaledCellDpSize.toPx() },
-                        cellWindow = cellWindow,
-                        modifier = Modifier.fillMaxSize(),
-                    )
-                    )
+                SelectionBoxOverlay(
+                    selectionSessionState = targetSelectionSessionState as SessionValue<SelectionState.Selection>,
+                    setSelectionState = selectionSessionStateValueHolder::setValue,
+                    getSelectionCellState = {
+                        getSelectionCellState(targetSelectionState)
+                    },
+                    scaledCellPixelSize = with(LocalDensity.current) { scaledCellDpSize.toPx() },
+                    cellWindow = cellWindow,
+                    modifier = Modifier.fillMaxSize(),
+                )
             }
         }
     }
 }
 
-/**
- * Computes the initial handles for a [SelectionState.SelectingBox.FixedSelectingBox].
- *
- * If there is a [SelectionState.SelectingBox.FixedSelectingBox.previousTransientSelectingBox], then these handles
- * will follow those offsets (to be transiently adjusted back to the rounded values).
- *
- * Otherwise, it will just be the direct fixed values.
- */
-private val SelectionState.SelectingBox.FixedSelectingBox.initialHandles get(): List<Offset> {
-    val initialHandleAOffset: Offset
-    val initialHandleBOffset: Offset
-    val initialHandleCOffset: Offset
-    val initialHandleDOffset: Offset
-
-    if (previousTransientSelectingBox != null) {
-        initialHandleAOffset = previousTransientSelectingBox.rect.topLeft
-        initialHandleBOffset = previousTransientSelectingBox.rect.topRight
-        initialHandleCOffset = previousTransientSelectingBox.rect.bottomRight
-        initialHandleDOffset = previousTransientSelectingBox.rect.bottomLeft
-    } else {
-        initialHandleAOffset = topLeft.toOffset()
-        initialHandleBOffset = (topLeft + IntOffset(width, 0)).toOffset()
-        initialHandleCOffset = (topLeft + IntOffset(width, height)).toOffset()
-        initialHandleDOffset = (topLeft + IntOffset(0, height)).toOffset()
-    }
-
-    return listOf(
-        initialHandleAOffset,
-        initialHandleBOffset,
-        initialHandleCOffset,
-        initialHandleDOffset,
+private fun ContentDrawScope.drawDropPreview(
+    dropPreview: CellStateDropState.DropPreview,
+    scaledCellDpSize: Dp,
+    cellStateOutlineColor: Color,
+) {
+    drawDashedRect(
+        selectionColor = cellStateOutlineColor,
+        strokeWidth = 2.dp.toPx(),
+        intervals = floatArrayOf(
+            24.dp.toPx(),
+            12.dp.toPx(),
+        ),
+        phase = 12.dp.toPx(),
+        rect = (
+            (
+                dropPreview.cellState.boundingBox.size.toIntRect().toRect()
+                    .translate(-dropPreview.cellState.boundingBox.size.toSize().center)
+                ) * scaledCellDpSize.toPx()
+            ).translate(dropPreview.offset),
     )
-}
-
-/**
- * The overlay for a [SelectionState.SelectingBox] [selectionState].
- *
- * This includes the selection box, along with 4 handles draggable in two dimensions to allow changing the selecting
- * box bounds.
- */
-@Suppress("LongMethod", "CyclomaticComplexMethod", "LongParameterList")
-@Composable
-private fun FixedSelectingBoxOverlay(
-    selectionSessionState: SessionValue<SelectionState.SelectingBox.FixedSelectingBox>,
-    setSelectionState: (SelectionState) -> Unit,
-    getSelectionCellState: () -> CellState,
-    scaledCellPixelSize: Float,
-    cellWindow: CellWindow,
-    modifier: Modifier = Modifier,
-) {
-    Box(modifier) {
-        /**
-         * The initial handles to initialize the handles with.
-         */
-        val initialHandles = selectionSessionState.value.initialHandles
-
-        /**
-         * The [DraggableAnchors2D] aligned to the current grid.
-         */
-        val handleAnchors = remember(scaledCellPixelSize, cellWindow) {
-            GridDraggableAnchors2d(scaledCellPixelSize, cellWindow)
-        }
-
-        /**
-         * State holders for the value change confirmation lambdas.
-         *
-         * These are initialized with a placeholder method, since this depends on the state of the other handles.
-         */
-        val confirmValueChangeStates = List(initialHandles.size) { index ->
-            key(index) {
-                remember { mutableStateOf({ _: IntOffset -> true }) }
-            }
-        }
-
-        val coroutineScope = rememberCoroutineScope()
-
-        /**
-         * A list of [Animatable]s for each handle representing the fractional part of the initial handle value, in
-         * cell coordinates.
-         *
-         * This will be initially added to the offset calculations, and animated to zero.
-         */
-        val transientSelectingBoxAnimatables = initialHandles.mapIndexed { index, offset ->
-            key(index) {
-                remember {
-                    Animatable(
-                        initialValue = offset - offset.round().toOffset(),
-                        typeConverter = Offset.VectorConverter,
-                        visibilityThreshold = Offset.VisibilityThreshold / scaledCellPixelSize,
-                    )
-                }
-            }
-        }
-
-        // Resolve the transient offsets to zero
-        transientSelectingBoxAnimatables.forEachIndexed { index, animatable ->
-            key(index) {
-                LaunchedEffect(animatable) {
-                    animatable.animateTo(Offset.Zero)
-                }
-            }
-        }
-
-        @Stable
-        class SelectionDraggableHandleState(
-            val state: HandleState,
-            transientSelectingBoxAnimatable: Animatable<Offset, AnimationVector2D>,
-            val horizontalPairState: HandleState,
-            val verticalPairState: HandleState,
-            val oppositeCornerState: HandleState,
-        ) {
-            val confirmValueChange: (IntOffset) -> Boolean = { intOffset ->
-                if (state.reentrancyCount == 0) {
-                    val minX = min(intOffset.x, oppositeCornerState.state.targetValue.x)
-                    val maxX = max(intOffset.x, oppositeCornerState.state.targetValue.x)
-                    val minY = min(intOffset.y, oppositeCornerState.state.targetValue.y)
-                    val maxY = max(intOffset.y, oppositeCornerState.state.targetValue.y)
-
-                    setSelectionState(
-                        SelectionState.SelectingBox.FixedSelectingBox(
-                            topLeft = IntOffset(minX, minY),
-                            width = maxX - minX,
-                            height = maxY - minY,
-                            previousTransientSelectingBox = null,
-                        ),
-                    )
-
-                    coroutineScope.launch {
-                        try {
-                            horizontalPairState.reentrancyCount++
-                            horizontalPairState.state.snapTo(
-                                IntOffset(
-                                    x = horizontalPairState.state.targetValue.x,
-                                    y = intOffset.y,
-                                ),
-                            )
-                        } finally {
-                            horizontalPairState.reentrancyCount--
-                        }
-                    }
-                    coroutineScope.launch {
-                        try {
-                            verticalPairState.reentrancyCount++
-                            verticalPairState.state.snapTo(
-                                IntOffset(
-                                    x = intOffset.x,
-                                    y = verticalPairState.state.targetValue.y,
-                                ),
-                            )
-                        } finally {
-                            verticalPairState.reentrancyCount--
-                        }
-                    }
-                }
-                true
-            }
-
-            val offsetCalculator: () -> Offset = {
-                val xReferenceState = when {
-                    state.state.isDraggingOrAnimating() -> state
-                    verticalPairState.state.isDraggingOrAnimating() -> verticalPairState
-                    else -> state
-                }
-                val yReferenceState = when {
-                    state.state.isDraggingOrAnimating() -> state
-                    horizontalPairState.state.isDraggingOrAnimating() -> horizontalPairState
-                    else -> state
-                }
-
-                Offset(
-                    xReferenceState.state.requireOffset().x,
-                    yReferenceState.state.requireOffset().y,
-                ) + transientSelectingBoxAnimatable.value * scaledCellPixelSize
-            }
-        }
-
-        val handleAnchoredDraggable2DStates =
-            initialHandles.mapIndexed { index, initialHandleOffset ->
-                key(index, scaledCellPixelSize, cellWindow) {
-                    rememberSaveable(
-                        saver = Saver(
-                            save = { packInts(it.currentValue.x, it.currentValue.y) },
-                            restore = {
-                                AnchoredDraggable2DState(
-                                    initialValue = IntOffset(unpackInt1(it), unpackInt2(it)),
-                                    animationSpec = spring(),
-                                    confirmValueChange = { intOffset ->
-                                        confirmValueChangeStates[index].value.invoke(intOffset)
-                                    },
-                                )
-                            },
-                        ),
-                    ) {
-                        AnchoredDraggable2DState(
-                            initialValue = initialHandleOffset.round(),
-                            animationSpec = spring(),
-                            confirmValueChange = { intOffset ->
-                                confirmValueChangeStates[index].value.invoke(intOffset)
-                            },
-                        )
-                    }.apply {
-                        updateAnchors(
-                            newAnchors = handleAnchors,
-                            newTarget = targetValue,
-                        )
-                    }
-                }
-            }
-
-        val handleStates = handleAnchoredDraggable2DStates.mapIndexed { index, handleAnchoredDraggable2DState ->
-            key(index) {
-                remember(handleAnchoredDraggable2DState) {
-                    HandleState(handleAnchoredDraggable2DState)
-                }
-            }
-        }
-
-        val handleAState = handleStates[0]
-        val handleBState = handleStates[1]
-        val handleCState = handleStates[2]
-        val handleDState = handleStates[3]
-
-        val selectionHandleStates = remember(handleStates, transientSelectingBoxAnimatables) {
-            listOf(
-                SelectionDraggableHandleState(
-                    state = handleAState,
-                    transientSelectingBoxAnimatable = transientSelectingBoxAnimatables[0],
-                    horizontalPairState = handleBState,
-                    verticalPairState = handleDState,
-                    oppositeCornerState = handleCState,
-                ),
-                SelectionDraggableHandleState(
-                    state = handleBState,
-                    transientSelectingBoxAnimatable = transientSelectingBoxAnimatables[1],
-                    horizontalPairState = handleAState,
-                    verticalPairState = handleCState,
-                    oppositeCornerState = handleDState,
-                ),
-                SelectionDraggableHandleState(
-                    state = handleCState,
-                    transientSelectingBoxAnimatable = transientSelectingBoxAnimatables[2],
-                    horizontalPairState = handleDState,
-                    verticalPairState = handleBState,
-                    oppositeCornerState = handleAState,
-                ),
-                SelectionDraggableHandleState(
-                    state = handleDState,
-                    transientSelectingBoxAnimatable = transientSelectingBoxAnimatables[3],
-                    horizontalPairState = handleCState,
-                    verticalPairState = handleAState,
-                    oppositeCornerState = handleBState,
-                ),
-            )
-        }
-
-        selectionHandleStates.forEachIndexed { index, selectionDraggableHandleState ->
-            confirmValueChangeStates[index].value = selectionDraggableHandleState.confirmValueChange
-        }
-
-        SelectingBox(
-            modifier = Modifier
-                .fillMaxSize()
-                .boxLayoutByHandles(
-                    handleAOffsetCalculator = selectionHandleStates[0].offsetCalculator,
-                    handleBOffsetCalculator = selectionHandleStates[1].offsetCalculator,
-                    handleCOffsetCalculator = selectionHandleStates[2].offsetCalculator,
-                    handleDOffsetCalculator = selectionHandleStates[3].offsetCalculator,
-                )
-                .cellStateDragAndDropSource(getSelectionCellState),
-        )
-
-        val parameterizedStringResolver = parameterizedStringResolver()
-
-        selectionHandleStates
-            .forEachIndexed { index, selectionDraggableHandleState ->
-                key(index) {
-                    val interactionSource = remember { MutableInteractionSource() }
-
-                    val isDragged by interactionSource.collectIsDraggedAsState()
-                    val isHovered by interactionSource.collectIsHoveredAsState()
-                    val isPressed by interactionSource.collectIsPressedAsState()
-
-                    val isActive = isDragged || isHovered || isPressed
-
-                    SelectionHandle(
-                        isActive = isActive,
-                        modifier = Modifier
-                            .offset {
-                                selectionDraggableHandleState.offsetCalculator().round()
-                            }
-                            .graphicsLayer {
-                                translationX = -size.width / 2f
-                                translationY = -size.height / 2f
-                            }
-                            .anchoredDraggable2D(
-                                state = selectionDraggableHandleState.state.state,
-                                interactionSource = interactionSource,
-                            )
-                            .semantics {
-                                val targetValue = selectionDraggableHandleState.state.state.targetValue
-                                contentDescription = parameterizedStringResolver(
-                                    Strings.SelectingBoxHandle(
-                                        targetValue.x,
-                                        targetValue.y,
-                                    ),
-                                )
-                            },
-                    )
-                }
-            }
-    }
-}
-
-@Stable
-private class HandleState(
-    val state: AnchoredDraggable2DState<IntOffset>,
-) {
-    var reentrancyCount by mutableStateOf(0)
-}
-
-@Composable
-private fun TransientSelectingBoxOverlay(
-    selectionState: SelectionState.SelectingBox.TransientSelectingBox,
-    scaledCellPixelSize: Float,
-    cellWindow: CellWindow,
-    modifier: Modifier = Modifier,
-) {
-    Box(modifier) {
-        val selectionRect = selectionState.rect.translate(-cellWindow.topLeft.toOffset())
-
-        val handleAOffsetCalculator = { selectionRect.topLeft * scaledCellPixelSize }
-        val handleBOffsetCalculator = { selectionRect.topRight * scaledCellPixelSize }
-        val handleCOffsetCalculator = { selectionRect.bottomRight * scaledCellPixelSize }
-        val handleDOffsetCalculator = { selectionRect.bottomLeft * scaledCellPixelSize }
-
-        val handleOffsetCalculators = listOf(
-            handleAOffsetCalculator,
-            handleBOffsetCalculator,
-            handleCOffsetCalculator,
-            handleDOffsetCalculator,
-        )
-
-        SelectingBox(
-            modifier = Modifier
-                .fillMaxSize()
-                .boxLayoutByHandles(
-                    handleAOffsetCalculator = {
-                        selectionRect.topLeft * scaledCellPixelSize
-                    },
-                    handleBOffsetCalculator = {
-                        selectionRect.topRight * scaledCellPixelSize
-                    },
-                    handleCOffsetCalculator = {
-                        selectionRect.bottomRight * scaledCellPixelSize
-                    },
-                    handleDOffsetCalculator = {
-                        selectionRect.bottomLeft * scaledCellPixelSize
-                    },
-                ),
-        )
-
-        handleOffsetCalculators.mapIndexed { index, offsetCalculator ->
-            key(index) {
-                SelectionHandle(
-                    isActive = index == 2,
-                    modifier = Modifier
-                        .offset {
-                            offsetCalculator().round()
-                        }
-                        .graphicsLayer {
-                            translationX = -size.width / 2f
-                            translationY = -size.height / 2f
-                        },
-                )
-            }
-        }
-    }
-}
-
-/**
- * A selection handle.
- */
-@Composable
-fun SelectionHandle(
-    isActive: Boolean,
-    modifier: Modifier = Modifier,
-    selectionColor: Color = MaterialTheme.colorScheme.secondary,
-) {
-    Box(
-        modifier = modifier.size(48.dp),
-        contentAlignment = Alignment.Center,
-    ) {
-        val size by animateDpAsState(if (isActive) 24.dp else 16.dp)
-        val elevation by animateDpAsState(if (isActive) 4.dp else 0.dp)
-        Surface(
-            modifier = Modifier.size(size),
-            shape = CircleShape,
-            shadowElevation = elevation,
-            color = selectionColor,
-        ) {}
-    }
 }
 
 /**
@@ -646,199 +273,6 @@ fun Modifier.boxLayoutByHandles(
     layout(constraints.maxWidth, constraints.maxHeight) {
         placeable.place(topLeft.round())
     }
-}
-
-/**
- * The selecting box itself.
- */
-@Composable
-fun SelectingBox(
-    // noinspection ComposeModifierWithoutDefault
-    modifier: Modifier,
-    selectionColor: Color = MaterialTheme.colorScheme.secondary,
-) {
-    Canvas(
-        modifier = modifier.fillMaxSize(),
-    ) {
-        drawRect(
-            color = selectionColor,
-            alpha = 0.2f,
-        )
-        drawSelectionRect(
-            selectionColor = selectionColor,
-            strokeWidth = 2.dp.toPx(),
-            intervals = floatArrayOf(
-                24.dp.toPx(),
-                24.dp.toPx(),
-            ),
-            phase = 12.dp.toPx(),
-        )
-    }
-}
-
-/**
- * Draws a selection rectangle, with the given [PathEffect].
- *
- * This results in the [PathEffect] being applied symmetrically via 8 line segments: from the middle of each side,
- * to each corner.
- */
-private fun DrawScope.drawSelectionRect(
-    selectionColor: Color,
-    strokeWidth: Float,
-    intervals: FloatArray,
-    phase: Float = 0f,
-    rect: Rect = size.toRect(),
-) {
-    // Draw the selection outlines in a way that is symmetric, and hides the extra of the lines near the
-    // corners with the selection handles
-    listOf(
-        rect.topCenter to rect.topLeft,
-        rect.topCenter to rect.topRight,
-        rect.centerLeft to rect.topLeft,
-        rect.centerLeft to rect.bottomLeft,
-        rect.bottomCenter to rect.bottomLeft,
-        rect.bottomCenter to rect.bottomRight,
-        rect.centerRight to rect.topRight,
-        rect.centerRight to rect.bottomRight,
-    ).forEach { (start, end) ->
-        drawDashedLine(
-            color = selectionColor,
-            start = start,
-            end = end,
-            strokeWidth = strokeWidth,
-            intervals = intervals,
-            phase = phase,
-        )
-    }
-}
-
-@Suppress("LongParameterList")
-internal expect fun DrawScope.drawDashedLine(
-    color: Color,
-    start: Offset,
-    end: Offset,
-    intervals: FloatArray,
-    phase: Float = 0f,
-    strokeWidth: Float = Stroke.HairlineWidth,
-    cap: StrokeCap = Stroke.DefaultCap,
-    @FloatRange(from = 0.0, to = 1.0) alpha: Float = 1.0f,
-    colorFilter: ColorFilter? = null,
-    blendMode: BlendMode = DefaultBlendMode,
-)
-
-data class GridDraggableAnchors2d(
-    private val scaledCellPixelSize: Float,
-    private val cellWindow: CellWindow,
-) : DraggableAnchors2D<IntOffset> {
-    override fun positionOf(anchor: IntOffset): Offset =
-        (anchor.toOffset() - cellWindow.topLeft.toOffset()) * scaledCellPixelSize
-
-    override fun hasPositionFor(anchor: IntOffset): Boolean = true
-    override fun closestAnchor(position: Offset): IntOffset =
-        (position / scaledCellPixelSize).round() + cellWindow.topLeft
-
-    override val size: Int = Int.MAX_VALUE
-}
-
-@Suppress("LongParameterList", "LongMethod")
-@Composable
-private fun SelectionBoxOverlay(
-    selectionSessionState: SessionValue<SelectionState.Selection>,
-    setSelectionState: (SelectionState) -> Unit,
-    getSelectionCellState: () -> CellState,
-    scaledCellPixelSize: Float,
-    cellWindow: CellWindow,
-    modifier: Modifier = Modifier,
-) {
-    val handleAnchors = remember(scaledCellPixelSize, cellWindow) {
-        GridDraggableAnchors2d(scaledCellPixelSize, cellWindow)
-    }
-
-    val initialOffset = selectionSessionState.value.offset
-
-    val currentSelectionSessionState by rememberUpdatedState(selectionSessionState)
-    val currentSetSelectionState by rememberUpdatedState(setSelectionState)
-
-    val confirmValueChange = { intOffset: IntOffset ->
-        setSelectionState(
-            currentSelectionSessionState.value.copy(
-                offset = intOffset,
-            ),
-        )
-        true
-    }
-
-    val draggable2DState = rememberSaveable(
-        saver = Saver(
-            save = { packInts(it.currentValue.x, it.currentValue.y) },
-            restore = {
-                AnchoredDraggable2DState(
-                    initialValue = IntOffset(unpackInt1(it), unpackInt2(it)),
-                    animationSpec = spring(),
-                    confirmValueChange = confirmValueChange,
-                )
-            },
-        ),
-    ) {
-        AnchoredDraggable2DState(
-            initialValue = initialOffset,
-            animationSpec = spring(),
-            confirmValueChange = confirmValueChange,
-        )
-    }.apply {
-        updateAnchors(
-            newAnchors = handleAnchors,
-            // Ensure the target value remains the same due to updating the anchors.
-            // This keeps the selection box stationary relative to the overall universe.
-            newTarget = targetValue,
-        )
-    }
-
-    // As a side-effect of dragging, update the underlying selection state to match the intermediate selection.
-    LaunchedEffect(draggable2DState) {
-        snapshotFlow { draggable2DState.currentValue }
-            .onEach { intOffset ->
-                currentSetSelectionState(
-                    currentSelectionSessionState.value.copy(
-                        offset = intOffset,
-                    ),
-                )
-            }
-            .collect()
-    }
-
-    val boundingBox = selectionSessionState.value.cellState.boundingBox
-
-    SelectingBox(
-        modifier = modifier
-            .fillMaxSize()
-            .boxLayoutByHandles(
-                handleAOffsetCalculator = {
-                    draggable2DState.requireOffset()
-                },
-                handleBOffsetCalculator = {
-                    draggable2DState.requireOffset() + Offset(
-                        boundingBox.width.toFloat(),
-                        0f,
-                    ) * scaledCellPixelSize
-                },
-                handleCOffsetCalculator = {
-                    draggable2DState.requireOffset() + Offset(
-                        boundingBox.width.toFloat(),
-                        boundingBox.height.toFloat(),
-                    ) * scaledCellPixelSize
-                },
-                handleDOffsetCalculator = {
-                    draggable2DState.requireOffset() + Offset(
-                        0f,
-                        boundingBox.height.toFloat(),
-                    ) * scaledCellPixelSize
-                },
-            )
-            .cellStateDragAndDropSource(getSelectionCellState)
-            .anchoredDraggable2D(draggable2DState),
-        selectionColor = MaterialTheme.colorScheme.tertiary,
-    )
 }
 
 fun <T> AnchoredDraggable2DState<T>.isDraggingOrAnimating(): Boolean =
